p336 중복 코드를 부모 클래스로 올려라
