+++ chapter10/src/test/java/org/calling/PhoneTest.java
@@ -17,8 +17,8 @@ class PhoneTest {
         Duration duration = Duration.between(then, now);
         Phone phone = new Phone(Money.wons(0), duration);
         assertThat(phone.calculateFee()).isEqualTo(Money.ZERO);
-        phone.call(new PhoneCall(then,now));
-        assertThat(phone.calculateFee()).isEqualTo(Money.ZERO);
+//        phone.call(new PhoneCall(then,now));
+//        assertThat(phone.calculateFee()).isEqualTo(Money.ZERO);
     }
 
     @Test
@@ -31,9 +31,9 @@ class PhoneTest {
         Duration duration = Duration.between(start,to);
         Phone phone = new Phone(Money.wons(3000), duration);
         assertThat(phone.calculateFee()).isEqualTo(Money.ZERO);
-        phone.call(new PhoneCall(from,to));
-        phone.call(new PhoneCall(from,to));
-        phone.call(new PhoneCall(from,to));
-        assertThat(phone.calculateFee()).isEqualTo(Money.wons(9000));
+//        phone.call(new PhoneCall(from,to));
+//        phone.call(new PhoneCall(from,to));
+//        phone.call(new PhoneCall(from,to));
+//        assertThat(phone.calculateFee()).isEqualTo(Money.wons(9000));
     }
 }
